Refactor statisticReport component
